op-cancel-appointment - move encrypted-bark-appointment into its own model file
